backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
python-multipart
python-dotenv
passlib[bcrypt]
python-jose
email-validator
pytest
httpx
